keno_ai_project_2/requirements.txt
streamlit
pandas
easyocr
tensorflow
pillow
